Inventory trpc

+++ src/server/api/routers/purchase.ts
@@ -29,6 +29,16 @@ export const purchaseRouter = createTRPCRouter({
                        price: parseFloat(input.price)
                     },
                 });
+                await ctx.prisma.book.update({
+                  where: {
+                    isbn: input.isbn
+                  },
+                  data:{
+                    inventory: {
+                      increment: parseInt(input.quantity)
+                    }
+                  }
+                })
             }
             else{
                 console.log("No purchase order under that ID")
@@ -50,18 +60,43 @@ export const purchaseRouter = createTRPCRouter({
       )
       .mutation(async ({ ctx, input }) => {
         try {
-          await ctx.prisma.purchase.update({
-            where:
-          {
-            id: input.id
-        },
-          data: {
-            purchaseOrderId: input.purchaseOrderId,
-            bookId: input.isbn,
-            quantity: parseInt(input.quantity),
-            price: parseFloat(input.price)
-          },
+          const purchase = await ctx.prisma.purchase.findFirst({
+              where:
+            {
+              id: input.id
+            }
           });
+          const book = await ctx.prisma.book.findFirst({
+            where:{
+              isbn: input.isbn
+            }
+          })
+          const change: number = parseInt(input.quantity) - purchase.quantity 
+          if (book.inventory + change >= 0){
+            await ctx.prisma.book.update({
+              where:{
+                isbn: book.isbn
+              },
+              data:{
+                inventory: book.inventory + change
+              }
+            })
+            await ctx.prisma.purchase.update({
+              where:
+            {
+              id: input.id
+          },
+            data: {
+              purchaseOrderId: input.purchaseOrderId,
+              bookId: input.isbn,
+              quantity: parseInt(input.quantity),
+              price: parseFloat(input.price)
+            },
+            });
+          }
+          else{
+            console.log("This change would make the inventory negative")
+          }
         } catch (error) {
           console.log(error);
         }
@@ -75,11 +110,40 @@ export const purchaseRouter = createTRPCRouter({
       )
       .mutation(async ({ ctx, input }) => {
         try {
-          await ctx.prisma.purchase.delete({
+          const purchase = await ctx.prisma.purchase.findFirst({
             where: {
               id: input.id
             }
           })
+          if (purchase){
+            const book = await ctx.prisma.book.findFirst({
+              where:{
+                isbn: purchase.bookId
+              }
+            })
+            const inventory: number = parseInt(book.inventory) - parseInt(purchase.quantity)
+            if (inventory >= 0){
+              console.log("Inventory: "+inventory)
+              await ctx.prisma.purchase.delete({
+                where: {
+                  id: input.id
+                }
+              })
+              await ctx.prisma.book.update({
+                where:{
+                  isbn: purchase.bookId
+                },
+                data:{
+                  inventory: inventory
+                }
+              })
+            }
+            else{
+              console.log("Deleting this purchase makes the inventory negative")
+            }
+          }
+          
+          
         } catch (error) {
           console.log(error);
         }
